backend/requirements.railway.txt
# Free deployment requirements (optimized for Railway free tier)
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.5.0
# Use CPU-only TensorFlow for smaller size
tensorflow-cpu==2.13.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
aiofiles==23.2.1